Separate cookie management functions
